Move App file in components folder
