+++ src/components/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import BooksList from './containers/BooksList';
-import BooksForm from './containers/BooksForm';
+import BooksList from '../containers/BooksList';
+import BooksForm from '../containers/BooksForm';
 
 const App = () => (
   <div className="container mt-4">
