+++ src/index.js
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 
 import { createStore } from 'redux';
-import App from './App';
+import App from './components/App';
 import books from './reducers/books';
 
 import './assets/bootstrap.min.css';
